==> 5_stats/05_DAA.Rmd <==
---
title: "Arsenic - 05: Differential Abundance Analysis"
author: "Jan Waelchli"
geometry: margin=2cm
output:
  pdf_document:
    toc: yes
    toc_depth: 3
---
   
```{r setup, include=FALSE, echo=F, warning=F, message=F,}

##clear the object from memory
rm(list=ls())

#knitr settings
knitr::opts_chunk$set(echo=TRUE, fig.align="center")
options(tinytex.verbose = TRUE)

#set seed
set.seed(100)

## set source to file location
library(rstudioapi)
setwd(dirname(getActiveDocumentContext()$path))

## load functions
source("functions/functions.R")
functions(path="functions/")

## installs (if necessary) and loads libraries
libraries()

```

```{r paths, echo=F, message=F, warning=F}

# Import

#set wd to RDS files
setwd("interim/01_Import_Normalization")

#Import RDS files
bDESIGN <- readRDS("bDESIGN.RDS")
bDAT <- readRDS("bDAT.RDS")
bDAT_norm <- readRDS("bDAT_norm.RDS")
bTAX <- readRDS("bTAX.RDS")

fDESIGN <- readRDS("fDESIGN.RDS")
fDAT <- readRDS("fDAT.RDS")
fDAT_norm <- readRDS("fDAT_norm.RDS")
fTAX <- readRDS("fTAX.RDS")

level_cols_genotype <- readRDS("level_cols_genotype.RDS")
level_cols_arsenic <- readRDS("level_cols_arsenic.RDS")

bPHYSEQ_phyla_melt <- readRDS("../02_Taxa_Analysis/bPHYSEQ_phyla_melt.RDS")
fPHYSEQ_phyla_melt <- readRDS("../02_Taxa_Analysis/fPHYSEQ_phyla_melt.RDS")

```


\pagebreak

# Taxa Response

Is there a core of sensitive microbial taxa? We searched sensitive ASVs – ASVs being differential abundant between WT and bx1. We answer the following question in non-arsenic and arsenic conditions:

**Q1: Are there sensitive ASVs between control and WT and bx1 samples in non-arsenic and arsenic soil?**

## Method

We answered the question by using four different tools to measure differential abundances - aldex2, acombc, maaslin2 and metagenomeSeq - and predict ASVs to be different if they were detected by 2 or more tools.

```{r prepro, echo=F, message=F, warning=F, error=F}

preprocess <- function(DESIGN, DAT, sample_var){
  
  #preprocess (from ancom.R function file)
  #out_cut: detect outlayers
  #zero_cut: removes ASVs must be present in 90% of samples
  prepro <- feature_table_pre_process(feature_table=DAT, meta_data=DESIGN, 
                                       sample_var=sample_var, group_var=NULL, 
                                       out_cut=0.5, zero_cut=0.9,
                                       lib_cut=0, neg_lb=F)
  #overwrite
  feature_table <-  prepro$feature_table # Preprocessed feature table
  meta_data <-  prepro$meta_data # Preprocessed metadata
  return(list(feature_table, meta_data))
}

```

## Genotype*Arsenic Effect

We check for each ASV if it is sensitive or not. Then, we show how many ASVs has been changed between the genotypes and how much of the relative abundance belongs to those sensitive ASVs.
 
```{r compute DAA, echo=F, message=F, warning=F, error=F, results='hide'}

#BACTERIA
#CTRL
bDESIGN_CTRL <- bDESIGN[bDESIGN$arsenic=="CTRL",]
bDESIGN_CTRL$genotype <- as.character(bDESIGN_CTRL$genotype) #needs to be a character for the run_all_DAA function
bDAT_CTRL <- bDAT[,colnames(bDAT)%in%bDESIGN_CTRL$bac_FLD]
bDAT_norm_CTRL <- bDAT_norm[,colnames(bDAT_norm)%in%bDESIGN_CTRL$bac_FLD]

temp <- preprocess(bDESIGN_CTRL, bDAT_CTRL, "bac_FLD")
feature_table <- temp[[1]]; meta_data <- temp[[2]]; rm(temp)
#search differential abundance
temp <- run_all_DAA(feature_table, meta_data, group = "genotype", DAT_sample_names =  "bac_FLD", reference = "genotype,WT")
bsumm_CTRL <- temp[[1]]; rm(temp)
bsens_ASVs_CTRL <- bsumm_CTRL$ASV[bsumm_CTRL$score>=2]

#arsenic
bDESIGN_arsenic <- bDESIGN[bDESIGN$arsenic=="arsenic",]
bDESIGN_arsenic$genotype <- as.character(bDESIGN_arsenic$genotype) #needs to be a character for the run_all_DAA function
bDAT_arsenic <- bDAT[,colnames(bDAT)%in%bDESIGN_arsenic$bac_FLD]
bDAT_norm_arsenic <- bDAT_norm[,colnames(bDAT_norm)%in%bDESIGN_arsenic$bac_FLD]

temp <- preprocess(bDESIGN_arsenic, bDAT_arsenic, "bac_FLD")
feature_table <- temp[[1]]; meta_data <- temp[[2]]; rm(temp)
#search differential abundance
temp <- run_all_DAA(feature_table, meta_data, group = "genotype", DAT_sample_names =  "bac_FLD", reference = "genotype,bx1")
bsumm_arsenic <- temp[[1]]; rm(temp)
bsens_ASVs_arsenic <- bsumm_arsenic$ASV[bsumm_arsenic$score>=2]

#FUNGI
#CTRL
fDESIGN_CTRL <- fDESIGN[fDESIGN$arsenic=="CTRL",]
fDESIGN_CTRL$genotype <- as.character(fDESIGN_CTRL$genotype) #needs to be a character for the run_all_DAA function
fDAT_CTRL <- fDAT[,colnames(fDAT)%in%fDESIGN_CTRL$fun_FLD]
fDAT_norm_CTRL <- fDAT_norm[,colnames(fDAT_norm)%in%fDESIGN_CTRL$fun_FLD]

temp <- preprocess(fDESIGN_CTRL, fDAT_CTRL, "fun_FLD")
feature_table <- temp[[1]]; meta_data <- temp[[2]]; rm(temp)
#search differential abundance
temp <- run_all_DAA(feature_table, meta_data, group = "genotype", DAT_sample_names =  "fun_FLD", reference = "genotype,WT")
fsumm_CTRL <- temp[[1]]; rm(temp)
fsens_ASVs_CTRL <- fsumm_CTRL$ASV[fsumm_CTRL$score>=2]

#arsenic
fDESIGN_arsenic <- fDESIGN[fDESIGN$arsenic=="arsenic",]
fDESIGN_arsenic$genotype <- as.character(fDESIGN_arsenic$genotype) #needs to be a character for the run_all_DAA function
fDAT_arsenic <- fDAT[,colnames(fDAT)%in%fDESIGN_arsenic$fun_FLD]
fDAT_norm_arsenic <- fDAT_norm[,colnames(fDAT_norm)%in%fDESIGN_arsenic$fun_FLD]

temp <- preprocess(fDESIGN_arsenic, fDAT_arsenic, "fun_FLD")
feature_table <- temp[[1]]; meta_data <- temp[[2]]; rm(temp)
#search differential abundance
temp <- run_all_DAA(feature_table, meta_data, group = "genotype", DAT_sample_names =  "fun_FLD", reference = "genotype,bx1")
fsumm_arsenic <- temp[[1]]; rm(temp)
fsens_ASVs_arsenic <- fsumm_arsenic$ASV[fsumm_arsenic$score>=2]

```


```{r bac summarise DAA, echo=F, message=F, warning=F, error=F}

#CTRL

#relative abundance
bDAT_norm_CTRL <- as.data.frame(bDAT_norm_CTRL)
bDAT_norm_rel_CTRL <- (bDAT_norm_CTRL/colSums(bDAT_norm_CTRL)) * 100
bDAT_norm_rel_CTRL <- as.data.frame(bDAT_norm_rel_CTRL[ rowSums(bDAT_norm_rel_CTRL) > 0,])

#means & standard error
means <- aggregate(t(bDAT_norm_rel_CTRL), list(bDESIGN_CTRL$genotype), mean) %>%
  tibble::column_to_rownames("Group.1") %>%
  t() %>% as.data.frame() %>%
  tibble::rownames_to_column("ASV")

se <- aggregate(t(bDAT_norm_rel_CTRL), list(bDESIGN_CTRL$genotype), parameters::standard_error) %>%
  tibble::column_to_rownames("Group.1") %>%
  t() %>% as.data.frame() %>%
  tibble::rownames_to_column("ASV")

#sort by abundance of CTRL samples
order <- order(-means$WT)
means <- means[order,]
se <- se[order,]

bsummary_CTRL <- tidyr::pivot_longer(means, !(ASV), names_to="genotype", values_to="means")
bsummary_CTRL$se <- tidyr::pivot_longer(se, !(ASV), names_to="genotype", values_to="se") %>% dplyr::pull(se)
bsummary_CTRL$genotype <- factor(bsummary_CTRL$genotype, levels=c("WT", "bx1"))
bsummary_CTRL$ASV <- factor(bsummary_CTRL$ASV, levels=unique(bsummary_CTRL$ASV)) #keep the sorted order of ASVs

#get score
bsummary_CTRL$score <- 0
  for(a in unique(bsummary_CTRL$ASV)){
    score <- bsumm_CTRL %>% dplyr::filter(ASV == a) %>% dplyr::pull(score)
    if(length(score)==1){bsummary_CTRL$score[(bsummary_CTRL$ASV==a)] <- score}
  }

#add if different abundant (score >= 2)
bsummary_CTRL$sensitive <- ifelse(bsummary_CTRL$score >= 2, "sensitive", "not_sensitive")
bsummary_CTRL$sensitive <- as.factor(bsummary_CTRL$sensitive)

#add abundace in trt compared to ctrl
bsummary_CTRL$abu_trt <- "unchanged"
for(a in unique(bsummary_CTRL$ASV)){
  mean_ctrl <- bsummary_CTRL$means[(bsummary_CTRL$ASV==a) & (bsummary_CTRL$genotype=="WT")]
  mean_trt <- bsummary_CTRL$means[(bsummary_CTRL$ASV==a) & (bsummary_CTRL$genotype=="bx1")]
  abu_sens <- bsummary_CTRL$sensitive[(bsummary_CTRL$ASV==a) & (bsummary_CTRL$genotype=="bx1")]
  if(abu_sens == "sensitive"){
    trt_dir <- ifelse(mean_ctrl > mean_trt, "higher in WT", "lower in WT")
    bsummary_CTRL$abu_trt[(bsummary_CTRL$ASV==a)] <- trt_dir
    }
}

#add phyla and class
bsummary_CTRL$phylum <- NA
bsummary_CTRL$class <- NA
bsummary_CTRL$family <- NA
for(a in unique(bsummary_CTRL$ASV)){
  bsummary_CTRL$phylum[bsummary_CTRL$ASV == a] <- as.character(bTAX$phylum[bTAX$ASV_ID == a])
  bsummary_CTRL$class[bsummary_CTRL$ASV == a] <- as.character(bTAX$class[bTAX$ASV_ID == a])
  bsummary_CTRL$family[bsummary_CTRL$ASV == a] <- as.character(bTAX$family[bTAX$ASV_ID == a])
}


#arsenic

#relative abundance
bDAT_norm_arsenic <- as.data.frame(bDAT_norm_arsenic)
bDAT_norm_rel_arsenic <- (bDAT_norm_arsenic/colSums(bDAT_norm_arsenic)) * 100
bDAT_norm_rel_arsenic <- as.data.frame(bDAT_norm_rel_arsenic[ rowSums(bDAT_norm_rel_arsenic) > 0,])

#means & standard error
means <- aggregate(t(bDAT_norm_rel_arsenic), list(bDESIGN_arsenic$genotype), mean) %>%
  tibble::column_to_rownames("Group.1") %>%
  t() %>% as.data.frame() %>%
  tibble::rownames_to_column("ASV")

se <- aggregate(t(bDAT_norm_rel_arsenic), list(bDESIGN_arsenic$genotype), parameters::standard_error) %>%
  tibble::column_to_rownames("Group.1") %>%
  t() %>% as.data.frame() %>%
  tibble::rownames_to_column("ASV")

#sort by abundance of arsenic samples
order <- order(-means$WT)
means <- means[order,]
se <- se[order,]

bsummary_arsenic <- tidyr::pivot_longer(means, !(ASV), names_to="genotype", values_to="means")
bsummary_arsenic$se <- tidyr::pivot_longer(se, !(ASV), names_to="genotype", values_to="se") %>% dplyr::pull(se)
bsummary_arsenic$genotype <- factor(bsummary_arsenic$genotype, levels=c("WT", "bx1"))
bsummary_arsenic$ASV <- factor(bsummary_arsenic$ASV, levels=unique(bsummary_arsenic$ASV)) #keep the sorted order of ASVs

#get score
bsummary_arsenic$score <- 0
  for(a in unique(bsummary_arsenic$ASV)){
    score <- bsumm_arsenic %>% dplyr::filter(ASV == a) %>% dplyr::pull(score)
    if(length(score)==1){bsummary_arsenic$score[(bsummary_arsenic$ASV==a)] <- score}
  }

#add if different abundant (score >= 2)
bsummary_arsenic$sensitive <- ifelse(bsummary_arsenic$score >= 2, "sensitive", "not_sensitive")
bsummary_arsenic$sensitive <- as.factor(bsummary_arsenic$sensitive)

#add abundace in trt compared to ctrl
bsummary_arsenic$abu_trt <- "unchanged"
for(a in unique(bsummary_arsenic$ASV)){
  mean_ctrl <- bsummary_arsenic$means[(bsummary_arsenic$ASV==a) & (bsummary_arsenic$genotype=="WT")]
  mean_trt <- bsummary_arsenic$means[(bsummary_arsenic$ASV==a) & (bsummary_arsenic$genotype=="bx1")]
  abu_sens <- bsummary_arsenic$sensitive[(bsummary_arsenic$ASV==a) & (bsummary_arsenic$genotype=="bx1")]
  if(abu_sens == "sensitive"){
    trt_dir <- ifelse(mean_ctrl > mean_trt, "higher in WT", "lower in WT")
    bsummary_arsenic$abu_trt[(bsummary_arsenic$ASV==a)] <- trt_dir
    }
}

#add phyla and class
bsummary_arsenic$phylum <- NA
bsummary_arsenic$class <- NA
bsummary_arsenic$family <- NA
for(a in unique(bsummary_arsenic$ASV)){
  bsummary_arsenic$phylum[bsummary_arsenic$ASV == a] <- as.character(bTAX$phylum[bTAX$ASV_ID == a])
  bsummary_arsenic$class[bsummary_arsenic$ASV == a] <- as.character(bTAX$class[bTAX$ASV_ID == a])
  bsummary_arsenic$family[bsummary_arsenic$ASV == a] <- as.character(bTAX$family[bTAX$ASV_ID == a])
}

```

```{r fun summarise DAA, echo=F, message=F, warning=F, error=F}

#CTRL

#relative abundance
fDAT_norm_CTRL <- as.data.frame(fDAT_norm_CTRL)
fDAT_norm_rel_CTRL <- (fDAT_norm_CTRL/colSums(fDAT_norm_CTRL)) * 100
fDAT_norm_rel_CTRL <- as.data.frame(fDAT_norm_rel_CTRL[ rowSums(fDAT_norm_rel_CTRL) > 0,])

#means & standard error
means <- aggregate(t(fDAT_norm_rel_CTRL), list(fDESIGN_CTRL$genotype), mean) %>%
  tibble::column_to_rownames("Group.1") %>%
  t() %>% as.data.frame() %>%
  tibble::rownames_to_column("ASV")

se <- aggregate(t(fDAT_norm_rel_CTRL), list(fDESIGN_CTRL$genotype), parameters::standard_error) %>%
  tibble::column_to_rownames("Group.1") %>%
  t() %>% as.data.frame() %>%
  tibble::rownames_to_column("ASV")

#sort by abundance of CTRL samples
order <- order(-means$WT)
means <- means[order,]
se <- se[order,]

fsummary_CTRL <- tidyr::pivot_longer(means, !(ASV), names_to="genotype", values_to="means")
fsummary_CTRL$se <- tidyr::pivot_longer(se, !(ASV), names_to="genotype", values_to="se") %>% dplyr::pull(se)
fsummary_CTRL$genotype <- factor(fsummary_CTRL$genotype, levels=c("WT", "bx1"))
fsummary_CTRL$ASV <- factor(fsummary_CTRL$ASV, levels=unique(fsummary_CTRL$ASV)) #keep the sorted order of ASVs

#get score
fsummary_CTRL$score <- 0
  for(a in unique(fsummary_CTRL$ASV)){
    score <- fsumm_CTRL %>% dplyr::filter(ASV == a) %>% dplyr::pull(score)
    if(length(score)==1){fsummary_CTRL$score[(fsummary_CTRL$ASV==a)] <- score}
  }

#add if different abundant (score >= 2)
fsummary_CTRL$sensitive <- ifelse(fsummary_CTRL$score >= 2, "sensitive", "not_sensitive")
fsummary_CTRL$sensitive <- as.factor(fsummary_CTRL$sensitive)

#add abundace in trt compared to ctrl
fsummary_CTRL$abu_trt <- "unchanged"
for(a in unique(fsummary_CTRL$ASV)){
  mean_ctrl <- fsummary_CTRL$means[(fsummary_CTRL$ASV==a) & (fsummary_CTRL$genotype=="WT")]
  mean_trt <- fsummary_CTRL$means[(fsummary_CTRL$ASV==a) & (fsummary_CTRL$genotype=="bx1")]
  abu_sens <- fsummary_CTRL$sensitive[(fsummary_CTRL$ASV==a) & (fsummary_CTRL$genotype=="bx1")]
  if(abu_sens == "sensitive"){
    trt_dir <- ifelse(mean_ctrl > mean_trt, "higher in WT", "lower in WT")
    fsummary_CTRL$abu_trt[(fsummary_CTRL$ASV==a)] <- trt_dir
    }
}

#add phyla and class
fsummary_CTRL$phylum <- NA
fsummary_CTRL$class <- NA
fsummary_CTRL$family <- NA
for(a in unique(fsummary_CTRL$ASV)){
  fsummary_CTRL$phylum[fsummary_CTRL$ASV == a] <- as.character(fTAX$phylum[fTAX$ASV_ID == a])
  fsummary_CTRL$class[fsummary_CTRL$ASV == a] <- as.character(fTAX$class[fTAX$ASV_ID == a])
  fsummary_CTRL$family[fsummary_CTRL$ASV == a] <- as.character(fTAX$family[fTAX$ASV_ID == a])
}


#arsenic

#relative abundance
fDAT_norm_arsenic <- as.data.frame(fDAT_norm_arsenic)
fDAT_norm_rel_arsenic <- (fDAT_norm_arsenic/colSums(fDAT_norm_arsenic)) * 100
fDAT_norm_rel_arsenic <- as.data.frame(fDAT_norm_rel_arsenic[ rowSums(fDAT_norm_rel_arsenic) > 0,])

#means & standard error
means <- aggregate(t(fDAT_norm_rel_arsenic), list(fDESIGN_arsenic$genotype), mean) %>%
  tibble::column_to_rownames("Group.1") %>%
  t() %>% as.data.frame() %>%
  tibble::rownames_to_column("ASV")

se <- aggregate(t(fDAT_norm_rel_arsenic), list(fDESIGN_arsenic$genotype), parameters::standard_error) %>%
  tibble::column_to_rownames("Group.1") %>%
  t() %>% as.data.frame() %>%
  tibble::rownames_to_column("ASV")

#sort by abundance of arsenic samples
order <- order(-means$WT)
means <- means[order,]
se <- se[order,]

fsummary_arsenic <- tidyr::pivot_longer(means, !(ASV), names_to="genotype", values_to="means")
fsummary_arsenic$se <- tidyr::pivot_longer(se, !(ASV), names_to="genotype", values_to="se") %>% dplyr::pull(se)
fsummary_arsenic$genotype <- factor(fsummary_arsenic$genotype, levels=c("WT", "bx1"))
fsummary_arsenic$ASV <- factor(fsummary_arsenic$ASV, levels=unique(fsummary_arsenic$ASV)) #keep the sorted order of ASVs

#get score
fsummary_arsenic$score <- 0
  for(a in unique(fsummary_arsenic$ASV)){
    score <- fsumm_arsenic %>% dplyr::filter(ASV == a) %>% dplyr::pull(score)
    if(length(score)==1){fsummary_arsenic$score[(fsummary_arsenic$ASV==a)] <- score}
  }

#add if different abundant (score >= 2)
fsummary_arsenic$sensitive <- ifelse(fsummary_arsenic$score >= 2, "sensitive", "not_sensitive")
fsummary_arsenic$sensitive <- as.factor(fsummary_arsenic$sensitive)

#add abundace in trt compared to ctrl
fsummary_arsenic$abu_trt <- "unchanged"
for(a in unique(fsummary_arsenic$ASV)){
  mean_ctrl <- fsummary_arsenic$means[(fsummary_arsenic$ASV==a) & (fsummary_arsenic$genotype=="WT")]
  mean_trt <- fsummary_arsenic$means[(fsummary_arsenic$ASV==a) & (fsummary_arsenic$genotype=="bx1")]
  abu_sens <- fsummary_arsenic$sensitive[(fsummary_arsenic$ASV==a) & (fsummary_arsenic$genotype=="bx1")]
  if(abu_sens == "sensitive"){
    trt_dir <- ifelse(mean_ctrl > mean_trt, "higher in WT", "lower in WT")
    fsummary_arsenic$abu_trt[(fsummary_arsenic$ASV==a)] <- trt_dir
    }
}

#add phyla and class
fsummary_arsenic$phylum <- NA
fsummary_arsenic$class <- NA
fsummary_arsenic$family <- NA
for(a in unique(fsummary_arsenic$ASV)){
  fsummary_arsenic$phylum[fsummary_arsenic$ASV == a] <- as.character(fTAX$phylum[fTAX$ASV_ID == a])
  fsummary_arsenic$class[fsummary_arsenic$ASV == a] <- as.character(fTAX$class[fTAX$ASV_ID == a])
  fsummary_arsenic$family[fsummary_arsenic$ASV == a] <- as.character(fTAX$family[fTAX$ASV_ID == a])
}

```

```{r bac show DAA, echo=F, message=F, warning=F, error=F}

#pander how many ASVs are sensitive

#BACTERIA
#CTRL
bchanged_abu_CTRL <- sum(bsummary_CTRL$means[bsummary_CTRL$sensitive=="sensitive" & bsummary_CTRL$genotype=="WT"])
bchanged_abu_CTRL <- paste0(round(bchanged_abu_CTRL,2), "%")
bsens_CTRL <- table(bsummary_CTRL$abu_trt[bsummary_CTRL$ASV %in% bsumm_CTRL$ASV & bsummary_CTRL$genotype=="bx1"])
bsens_CTRL <- c(bsens_CTRL, "0"); names(bsens_CTRL)[2] <- "lower in WT" #add lower in WT=0
bsens_CTRL <- c(bsens_CTRL, "0"); names(bsens_CTRL)[3] <- "higher in WT" #add higher in WT=0
bsens_CTRL <- bsens_CTRL[c(2,1,3)]
#pander(bsens_CTRL, caption="Bacteria CTRL (no arsenic): genotype effect")

#sensitive ASVs
bsensitive_CTRL <- tidyr::pivot_wider(bsummary_CTRL,id_cols = !(c(se, score)), names_from = "genotype", values_from = c("means", "abu_trt", "sensitive"))
bsensitive_CTRL <- bsensitive_CTRL[bsensitive_CTRL$sensitive_bx1 =="sensitive",] #sensitive
bsensitive_CTRL <- bsensitive_CTRL[,colnames(bsensitive_CTRL) %in% c("ASV","means_WT","means_bx1", "phylum", "class")] #column to pander
bsensitive_CTRL$means_WT <- paste0(round(bsensitive_CTRL$means_WT,3),"%")
bsensitive_CTRL$means_bx1 <- paste0(round(bsensitive_CTRL$means_bx1,3),"%")
colnames(bsensitive_CTRL)[4:5] <- c("abu WT", "abu bx1")
#pander(bsensitive_CTRL, caption="Bacteria: sensitive ASVs in CTRL")

#arsenic
bchanged_abu_arsenic <- sum(bsummary_arsenic$means[bsummary_arsenic$sensitive=="sensitive" & bsummary_arsenic$genotype=="WT"])
bchanged_abu_arsenic <- paste0(round(bchanged_abu_arsenic,2), "%")
bsens_arsenic <- table(bsummary_arsenic$abu_trt[bsummary_arsenic$ASV %in% bsumm_arsenic$ASV & bsummary_arsenic$genotype=="bx1"])
bsens_arsenic <- c(bsens_arsenic, "0"); names(bsens_arsenic)[2] <- "lower in WT" #add lower in WT=0
bsens_arsenic <- c(bsens_arsenic, "0"); names(bsens_arsenic)[3] <- "higher in WT" #add higher in WT=0
bsens_arsenic <- bsens_arsenic[c(2,1,3)]
#pander(bsens_arsenic, caption="Bacteria arsenic: genotype effect")

#sensitive ASVs
bsensitive_arsenic <- tidyr::pivot_wider(bsummary_arsenic,id_cols = !(c(se, score)), names_from = "genotype", values_from = c("means", "abu_trt", "sensitive"))
bsensitive_arsenic <- bsensitive_arsenic[bsensitive_arsenic$sensitive_bx1 =="sensitive",] #sensitive
bsensitive_arsenic <- bsensitive_arsenic[,colnames(bsensitive_arsenic) %in% c("ASV","means_WT","means_bx1", "phylum", "class")] #column to pander
bsensitive_arsenic$means_WT <- paste0(round(bsensitive_arsenic$means_WT,3),"%")
bsensitive_arsenic$means_bx1 <- paste0(round(bsensitive_arsenic$means_bx1,3),"%")
colnames(bsensitive_arsenic)[4:5] <- c("abu WT", "abu bx1")
#pander(bsensitive_arsenic, caption="Bacteria: sensitive ASVs in arsenic")

#combine into one df
bsens <- rbind(bsens_CTRL, bsens_arsenic)
bsens <- cbind(taxa=c("bac","bac"), arsenic=c("CTRL","arsenic"), bsens)
rownames(bsens) <- NULL #remove rownames
#add total abu
abu_WT <- c(gsub("%", "", bsensitive_CTRL$`abu WT`) %>% as.numeric() %>% sum() %>% paste0("%"),
              gsub("%", "", bsensitive_arsenic$`abu WT`) %>% as.numeric() %>% sum() %>% paste0("%"))
bsens <- cbind(bsens, `rel. abu. of sens. ASVs`=abu_WT)
pander(bsens, caption = "Bacteria: genotype effect")

#write_excel_csv(as.data.frame(bsens), "figs_for_paper/bacteria_sensitive.csv")

```

`r bchanged_abu_CTRL` in non-arsenic and `r bchanged_abu_arsenic` in arsenic conditions of the bacterial community was changed in abundance due to genotype.

\vspace{5mm}

```{r fun show DAA, echo=F, message=F, warning=F, error=F}

#pander how many ASVs are sensitive

#FUNGI
#CTRL
fchanged_abu_CTRL <- sum(fsummary_CTRL$means[fsummary_CTRL$sensitive=="sensitive" & fsummary_CTRL$genotype=="WT"])
fchanged_abu_CTRL <- paste0(round(fchanged_abu_CTRL,2), "%")
fsens_CTRL <- table(fsummary_CTRL$abu_trt[fsummary_CTRL$ASV %in% fsumm_CTRL$ASV & fsummary_CTRL$genotype=="bx1"])
fsens_CTRL <- c(fsens_CTRL, "0"); names(fsens_CTRL)[3] <- "lower in WT" #add lower in WT=0
fsens_CTRL <- fsens_CTRL[c(3,2,1)]
#pander(fsens_CTRL, caption="Fungi CTRL (no arsenic): genotype effect")

#sensitive ASVs
fsensitive_CTRL <- tidyr::pivot_wider(fsummary_CTRL,id_cols = !(c(se, score)), names_from = "genotype", values_from = c("means", "abu_trt", "sensitive"))
fsensitive_CTRL <- fsensitive_CTRL[fsensitive_CTRL$sensitive_bx1 =="sensitive",] #sensitive
fsensitive_CTRL <- fsensitive_CTRL[,colnames(fsensitive_CTRL) %in% c("ASV","means_WT","means_bx1", "phylum", "class")] #column to pander
fsensitive_CTRL$means_WT <- paste0(round(fsensitive_CTRL$means_WT,3),"%")
fsensitive_CTRL$means_bx1 <- paste0(round(fsensitive_CTRL$means_bx1,3),"%")
colnames(fsensitive_CTRL)[4:5] <- c("abu WT", "abu bx1")
#pander(fsensitive_CTRL, caption="Fungi: sensitive ASVs in CTRL")

#arsenic
fchanged_abu_arsenic <- sum(fsummary_arsenic$means[fsummary_arsenic$sensitive=="sensitive" & fsummary_arsenic$genotype=="WT"])
fchanged_abu_arsenic <- paste0(round(fchanged_abu_arsenic,2), "%")
fsens_arsenic <- table(fsummary_arsenic$abu_trt[fsummary_arsenic$ASV %in% fsumm_arsenic$ASV & fsummary_arsenic$genotype=="bx1"])
fsens_arsenic <- c(fsens_arsenic, "0"); names(fsens_arsenic)[2] <- "lower in WT" #add lower in WT=0
fsens_arsenic <- c(fsens_arsenic, "0"); names(fsens_arsenic)[3] <- "higher in WT" #add higher in WT=0
fsens_arsenic <- fsens_arsenic[c(2,1,3)]
#pander(fsens_arsenic, caption="Fungi arsenic: genotype effect")

#sensitive ASVs
fsensitive_arsenic <- tidyr::pivot_wider(fsummary_arsenic,id_cols = !(c(se, score)), names_from = "genotype", values_from = c("means", "abu_trt", "sensitive"))
fsensitive_arsenic <- fsensitive_arsenic[fsensitive_arsenic$sensitive_bx1 =="sensitive",] #sensitive
fsensitive_arsenic <- fsensitive_arsenic[,colnames(fsensitive_arsenic) %in% c("ASV","means_WT","means_bx1", "phylum", "class")] #column to pander
fsensitive_arsenic$means_WT <- paste0(round(fsensitive_arsenic$means_WT,3),"%")
fsensitive_arsenic$means_bx1 <- paste0(round(fsensitive_arsenic$means_bx1,3),"%")
colnames(fsensitive_arsenic)[4:5] <- c("abu WT", "abu bx1")
#pander(fsensitive_arsenic, caption="Fungi: sensitive ASVs in arsenic")

#combine into one df
fsens <- rbind(fsens_CTRL, fsens_arsenic)
fsens <- cbind(taxa=c("fungi","fungi"), arsenic=c("CTRL","arsenic"), fsens)
rownames(fsens) <- NULL #remove rownames
#add total abu
abu_WT <- c(gsub("%", "", fsensitive_CTRL$`abu WT`) %>% as.numeric() %>% sum() %>% paste0("%"),
              gsub("%", "", fsensitive_arsenic$`abu WT`) %>% as.numeric() %>% sum() %>% paste0("%"))
fsens <- cbind(fsens, `rel. abu. of sens. ASVs`=abu_WT)
pander(fsens, caption = "Fungi: gneotype effect")

#write_excel_csv(as.data.frame(fsens), "figs_for_paper/fungi_sensitive.csv")

#remove automatically created DAA folder
unlink("DAA example/", recursive = T)

```

`r fchanged_abu_CTRL` in non-arsenic and `r fchanged_abu_arsenic` in arsenic conditions of the fungal community was changed in abundance due to genotype.

**Conclusion:** Most ASVs are insensitive.

\pagebreak
